Review and Checkout devs
implemented review and checkout microservices in the backend.

added product rating functionality.

improved PastOrders and CurrentOrders pages.

implemented pagination on the frontend.

used Pageable in the backend for comments.

created states for the frontend and DTOs for the backend.

created checkoutApiSlice and reviewApiSlice in the frontend with Redux.

created CheckoutPage in the frontend.

some bugs fixed.

and other devs..